backend/requirements_backend.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
pymongo==4.6.1
dnspython==2.7.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-storage==2.18.2
PyJWT==2.10.1
pydantic==2.11.7
pytest==8.4.2
waitress==3.0.0
paho-mqtt==2.1.0
